% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl_sub}
\alias{adsl_sub}
\title{CDISC ADSL Subsetting Data}
\format{
A \link{adsl_sub} data set contains 120 observations and 14 variables.
And the description of each variable has been labeled in data set.
}
\usage{
adsl_sub
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This ADSL is created by subsetting the CDISC ADSL with 60 subjects in each of two
treatments like placebo and Xanomeline (this one corresponds to high dose level
in the original ADSL).
}
\keyword{datasets}
