% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{BAsummary-class}
\alias{BAsummary-class}
\alias{BAsummary}
\title{BAsummary Class}
\usage{
BAsummary(call, data, stat, param)
}
\arguments{
\item{call}{(\code{call})\cr function call.}

\item{data}{(\code{data.frame})\cr stores the raw data from input.}

\item{stat}{(\code{list})\cr contains several statistics for numeric data.}

\item{param}{(\code{list})\cr list of relevant parameters.}
}
\value{
An object of class \code{BAsummary}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{BAsummary} class is used to display the BlandAltman analysis and outliers.
}
\section{Slots}{

\describe{
\item{\code{call}}{call}

\item{\code{data}}{data}

\item{\code{outlier}}{outlier}

\item{\code{param}}{param}
}}

