% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blandAltman.R
\name{blandAltman}
\alias{blandAltman}
\title{Calculate statistics for Bland-Altman}
\usage{
blandAltman(
  x,
  y,
  sid = NULL,
  type1 = 3,
  type2 = 5,
  outlier = FALSE,
  confint = 0.95
)
}
\arguments{
\item{x}{(\code{numeric})\cr reference method.}

\item{y}{(\code{numeric})\cr test method.}

\item{sid}{(\code{numeric} or \code{string}) sample id.}

\item{type1}{(\code{integer})\cr specifying a specific difference for absolute difference, default is 3.}

\item{type2}{(\code{integer})\cr specifying a specific difference for relative difference, default is 5.}

\item{outlier}{(\code{logic})\cr whether to calculate the outliers with 4E&4Er method.}

\item{confint}{(\code{numeric})\cr significance level for two side, default is 0.95.}
}
\value{
A object with \code{\link{BAsummary}} class that contains the BlandAltman analysis.
}
\description{
Calculate statistics for Bland-Altman
}
\examples{
data("platelet")
blandAltman(x = platelet$Comparative, y = platelet$Candidate)
data("creatinine", package = "mcr")
blandAltman(x = creatinine$serum.crea, y = creatinine$plasma.crea, outlier = TRUE)
}
\seealso{
\code{\link[=h_difference]{h_difference()}} to see the type details.
}
