% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calcium}
\alias{calcium}
\title{Reference Interval Data}
\format{
A \link{calcium} data set contains 240 observations and 3 variables.
\describe{
\item{Sample}{Sample id}
\item{Value}{Measurements from target subjects}
\item{Group}{Sex group of target subjects}
}
}
\source{
CLSI-EP28A3 Table 4. is cited in this data set.
}
\usage{
calcium
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This example \code{\link{calcium}} can be used to compute the reference range of
Calcium in 240 medical students by sex.
}
\keyword{datasets}
