% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{Desc-class}
\alias{Desc-class}
\alias{Desc}
\title{Descriptive Statistics Class}
\usage{
Desc(func, mat, stat)
}
\arguments{
\item{func}{(\code{character})\cr name of function.}

\item{mat}{(\code{data.frame})\cr intermediate data with long form, easy for post-processing.}

\item{stat}{(\code{data.frame})\cr final data with wide form, easy for presentation.}
}
\value{
An object of class \code{Desc}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{Desc} class serves as the store for results from frequency and univariate
statistics analysis.
}
\section{Slots}{

\describe{
\item{\code{func}}{func}

\item{\code{mat}}{mat}

\item{\code{stat}}{stat}
}}

