% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_stat.R
\name{diagTab}
\alias{diagTab}
\title{Creates Contingency Table}
\usage{
diagTab(formula = ~., data, ..., rlevels = NULL, clevels = NULL)
}
\arguments{
\item{formula}{(\code{numeric})\cr a \link{formula} object with the cross-classifying
variables (separated by \code{+}) on the right hand side. The row name of contingency
is represented by the variable to the left of the \code{+} sign, and the col name
is represented by the variable to the right.}

\item{data}{(\code{data.frame} or \code{matrix})\cr a data frame or matrix.}

\item{...}{other arguments to be passed to \code{\link[=xtabs]{xtabs()}}.}

\item{rlevels}{(\code{vector})\cr a character vector of known levels for row name
of contingency table.}

\item{clevels}{(\code{vector})\cr a character vector of known levels for col name
of contingency table.}
}
\value{
A object \code{matrix} contains the 2x2 contingency table.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates a 2x2 contingency table from the data frame or matrix for the qualitative
performance of downstream analysis.
}
\examples{
data("qualData")
qualData \%>\% diagTab(formula = ~ CandidateN + ComparativeN)
qualData \%>\%
  diagTab(
    formula = ~ CandidateN + ComparativeN,
    rlevels = c(1, 0), clevels = c(1, 0)
  )
}
\seealso{
\code{\link[=Summary]{Summary()}} for object to calculate diagnosis performance.
}
