% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dixon_outlier}
\alias{dixon_outlier}
\title{Detect Dixon Outlier}
\usage{
dixon_outlier(x)
}
\arguments{
\item{x}{(\code{numeric})\cr numeric input.}
}
\value{
A list contains outliers and vector without outliers.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Help function detects the potential outlier with Dixon method, following the
rules of EP28A3 and NMPA guideline for establishment of reference range.
}
\examples{
x <- c(13.6, 44.4, 45.9, 11.9, 41.9, 53.3, 44.7, 95.2, 44.1, 50.7, 45.2, 60.1, 89.1)
dixon_outlier(x)
}
