% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{esd.critical}
\alias{esd.critical}
\title{Compute the Critical Value for ESD Test}
\usage{
esd.critical(alpha, N, i)
}
\arguments{
\item{alpha}{(\code{numeric})\cr type-I-risk, \eqn{\alpha}.}

\item{N}{(\code{integer})\cr the total number of samples.}

\item{i}{(\code{integer})\cr the iteration number, less than the number of  biggest
potential outliers.}
}
\value{
a lambda value calculated from the formula.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A helper function to find the lambda for all potential outliers in each iteration.
}
\examples{
esd.critical(alpha = 0.05, N = 100, i = 1)
}
