% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Get Regression Coefficients}
\usage{
getCoefficients(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[mcr:MCResult.getCoefficients]{mcr::getCoefficients}}
  \describe{
    \item{\code{.Object}}{object of class "MCResult".}
  }}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A copy from \link[mcr:MCResult.getCoefficients]{mcr::getCoefficients} in \code{mcr} package
}
\examples{
data(platelet)
fit <- mcreg(
  x = platelet$Comparative, y = platelet$Candidate,
  method.reg = "Deming", method.ci = "jackknife"
)
getCoefficients(fit)
}
