% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glucose}
\alias{glucose}
\title{Inermediate Precision Data}
\format{
A \link{glucose} data set contains 80 observations and 3 variables.
\describe{
\item{day}{day number}
\item{run}{run number}
\item{value}{measurement value}
}
}
\source{
CLSI-EP05A3 Table A1. Glucose Precision Evaluation Measurements (mg/dL)
is cited in this data set.
}
\usage{
glucose
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This data set consists of the Glucose intermediate precision data in the CLSI
EP05-A3 guideline.
}
\references{
EP05A3: Evaluation of Precision of Quantitative Measurement Procedures.
}
\keyword{datasets}
