% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_fmt_range}
\alias{h_fmt_range}
\title{Format and Concatenate to Range}
\usage{
h_fmt_range(num1, num2, digits = c(2, 2), width = c(6, 6))
}
\arguments{
\item{num1}{(\code{numeric})\cr first numeric input.}

\item{num2}{(\code{numeric})\cr second numeric input.}

\item{digits}{(\code{integer})\cr the desired number of digits after the decimal point.}

\item{width}{(\code{integer})\cr the total field width.}
}
\value{
A single character.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Help function to format numeric data as strings and concatenate into a
single character range.
}
\examples{
h_fmt_range(num1 = 3.14, num2 = 3.14, width = c(4, 4))
}
\seealso{
\code{\link[=h_fmt_num]{h_fmt_num()}}
}
