% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_summarize}
\alias{h_summarize}
\title{Summarize the basic statistics}
\usage{
h_summarize(x, confint = 0.95)
}
\arguments{
\item{x}{(\code{numeric})\cr input numeric vector.}

\item{confint}{(\code{numeric})\cr significance level, default is 0.95.}
}
\value{
a verctor contains several statistics, such as n, mean, median, min,
max, q25, q75, sd, se, limit of agreement of limit and confidence interval .
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Help function summarizes the statistics as needed.
}
\examples{
h_summarize(1:50)
}
