% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{MCR-class}
\alias{MCR-class}
\alias{MCR}
\title{Method Comparison Regression Class}
\usage{
MCR(data, coef, mnames, regmeth)
}
\arguments{
\item{data}{(\code{data.frame})\cr original data.}

\item{coef}{(\code{numeric})\cr a numeric vector contains slope and intercept.}

\item{mnames}{(\code{character})\cr name of X and Y assays, default are 'Method1'
and 'Method2' if you have not defined them in \code{mcreg} function.}

\item{regmeth}{(\code{character})\cr name of regression.}
}
\value{
An object of class \code{MCR}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{MCR} class serves as a simplified version of \code{MCResult} from \code{mcr} package.
As the \code{mcr} package are not available in CRAN, this class is took as the temporary
replacement of it, which only contains the some necessaries for \code{autoplot}.
}
\section{Slots}{

\describe{
\item{\code{data}}{data}

\item{\code{coef}}{coef}

\item{\code{mnames}}{mnames}

\item{\code{regmeth}}{regmeth}
}}

