% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mcradds-package}
\alias{mcradds}
\alias{mcradds-package}
\title{\code{mcradss} Package}
\description{
\code{mcradss} facilitates analyzing and reporting in IVD trials.
}
\author{
\strong{Maintainer}: Kai Gu \email{gukai1212@163.com}

}
