% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{MCTab-class}
\alias{MCTab-class}
\alias{MCTab}
\title{MCTab Class}
\usage{
MCTab(tab, candidate, comparative)
}
\arguments{
\item{tab}{(\code{matrix})\cr \code{matrix} class converted from \code{\link[=xtabs]{xtabs()}} to display 2x2 contingency table.}

\item{candidate}{(\code{list})\cr candidate information, like measurements,
positive and negative levels.}

\item{comparative}{(\code{list})\cr comparative information, like measurements,
positive and negative levels.}
}
\value{
An object of class \code{MCTab}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{MCTab} class serves as the store for 2x2 contingency table, candidate and
comparative measurement's information.
}
\section{Slots}{

\describe{
\item{\code{slots}}{tab candidate comparative}
}}

