% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonparRanks}
\alias{nonparRanks}
\title{Nonparametric Rank Number of Reference Interval}
\format{
A \link{nonparRanks} data set contains 882 observations and 3 variables.
\describe{
\item{SampleSize}{sample size}
\item{Lower}{lower rank}
\item{Upper}{upper rank}
}
}
\source{
CLSI-EP28A3 Table 8. is cited in this data set.
}
\usage{
nonparRanks
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This data shows the rank number for computing the confidence interval of
nonparametric reference limit when the samples within 119-1000 values. But the
reference interval must be 95\% and the confidence interval is 90\%.
}
\references{
EP28-A3c: Defining, Establishing, and Verifying Reference Intervals
in the Clinical Laboratory.
}
\keyword{datasets}
