% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcr.R
\name{printSummary}
\alias{printSummary}
\title{Print Summary of a Regression Analysis}
\usage{
printSummary(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[mcr:MCResult.printSummary]{mcr::printSummary}}
  \describe{
    \item{\code{.Object}}{object of type "MCResult".}
  }}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A copy from \link[mcr:MCResult.printSummary]{mcr::printSummary} in \code{mcr} package
}
\examples{
data(platelet)
fit <- mcreg(x = platelet$Comparative, y = platelet$Candidate,
             method.reg = "Deming", method.ci = "jackknife")
printSummary(fit)
}
