% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qualData}
\alias{qualData}
\title{Example \code{qualData} Data}
\format{
A \link{qualData} contains 200 observations and 3 variables.
\describe{
\item{sample}{Sample id}
\item{Comparative}{Measurement from comparative analyzer with 1=positive and 0=negative}
\item{Candidate}{Measurement from candidate analyzer with 1=positive and 0=negative}
}
}
\usage{
qualData
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This simulated data \code{\link{qualData}} can be used to calculate the
quantitative performance such as sensitivity and specificity.
}
\seealso{
\link{platelet} which contains quantitative data comparing
platelet results from two analyzers.
}
\keyword{datasets}
