% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{RefInt-class}
\alias{RefInt-class}
\alias{RefInt}
\title{Reference Interval Class}
\usage{
RefInt(call, method, n, data, outlier, refInt, confInt)
}
\arguments{
\item{call}{(\code{call})\cr function call.}

\item{method}{(\code{character})\cr method names of reference interval and
confidence interval.}

\item{n}{(\code{numeric})\cr number of available samples.}

\item{data}{(\code{numeric})\cr numeric raw measurements, no outlier removed.}

\item{outlier}{(\code{list})\cr list of outliers that contains the index and number
of outliers, and the data without outliers.}

\item{refInt}{(\code{numeric})\cr number of reference interval.}

\item{confInt}{(\code{list})\cr list of the confidence interval of lower and upper
of reference limit.}
}
\value{
An object of class \code{RefInt}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{RefInt} class serves as the store for results in reference
Interval calculation.
}
\section{Slots}{

\describe{
\item{\code{call}}{call}

\item{\code{method}}{method}

\item{\code{n}}{n}

\item{\code{data}}{data}

\item{\code{outlier}}{outlier}

\item{\code{refInt}}{refInt}

\item{\code{confInt}}{confInt}
}}

