% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{SampleSize-class}
\alias{SampleSize-class}
\alias{SampleSize}
\title{SampleSize Class}
\usage{
SampleSize(call, method, n, param)
}
\arguments{
\item{call}{(\code{call})\cr function call.}

\item{method}{(\code{character})\cr method name.}

\item{n}{(\code{numeric})\cr number of sample size.}

\item{param}{(\code{list})\cr list of relevant parameters.}
}
\value{
An object of class \code{SampleSize}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{SampleSize} class serves as the store for results and parameters in sample
size calculation.
}
\section{Slots}{

\describe{
\item{\code{call}}{call}

\item{\code{method}}{method}

\item{\code{n}}{n}

\item{\code{param}}{param}
}}

