% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual_stat.R
\name{summary,MCTab-method}
\alias{summary,MCTab-method}
\title{Summary Method for \code{MCTab} Objects}
\usage{
\S4method{summary}{MCTab}(
  object,
  withref = TRUE,
  method = c("wilson", "wald", "clopper-pearson"),
  alpha = 0.05,
  ...
)
}
\arguments{
\item{object}{(\code{MCTab})\cr input from \link{diagTab} function to create 2x2 contingency table.}

\item{withref}{(\code{logical})\cr weather the comparison test is reference/gold standard or not.}

\item{method}{(\code{string})\cr string specifying the which method to use, \code{wilson} is default.
Options can be \code{wilson}, \code{wald} and \code{clopper-pearson}, see \link[DescTools:BinomCIn]{DescTools::BinomCIn}.}

\item{alpha}{(\code{numeric})\cr type-I-risk, \eqn{\alpha}.}

\item{...}{other arguments to be passed to \link[DescTools:BinomCIn]{DescTools::BinomCIn}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Provides a concise summary of the content of \code{\link{MCTab}} objects. Computes
sensitivity, specificity, positive and negative predictive values and positive
and negative likelihood ratios for a diagnostic test with reference/gold standard.
Computes positive and negative percent agreement, and overall percent agreement
when the new test is evaluated by comparison to a non-reference standard.
}
\examples{
summary(m, method = "wilson")
summary(m, method = "wilson", withref = FALSE)
}
