% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{tpROC-class}
\alias{tpROC-class}
\alias{tpROC}
\title{Test for Paired ROC Class}
\usage{
tpROC(testROC, refROC, method, H0, stat)
}
\arguments{
\item{testROC}{(\code{list})\cr object from \code{pRPC::roc()} function for test assay.}

\item{refROC}{(\code{list})\cr object from \code{pRPC::roc()} function for reference/standard assay.}

\item{method}{(\code{character})\cr method of hypothesis test.}

\item{H0}{(\code{numeric})\cr margin of test.}

\item{stat}{(\code{list})\cr list that contains the difference comparing results,
such as the difference of AUC, standard error, confidence interval, Z statistic
and P value.}
}
\value{
An object of class \code{tpROC}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{tpROC} class serves as the store for results in testing the AUC of paired
two-sample assays.
}
\section{Slots}{

\describe{
\item{\code{testROC}}{testROC}

\item{\code{refROC}}{refROC}

\item{\code{method}}{method}

\item{\code{stat}}{stat}
}}

