% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_methods.R, R/qual_stat.R
\name{getAccuracy}
\alias{getAccuracy}
\alias{getAccuracy,MCTab-method}
\title{Summary Method for \code{MCTab} Objects}
\usage{
getAccuracy(object, ...)

\S4method{getAccuracy}{MCTab}(
  object,
  withref = TRUE,
  method = c("wilson", "wald", "clopper-pearson"),
  alpha = 0.05,
  digits = 4,
  ...
)
}
\arguments{
\item{object}{(\code{MCTab})\cr input from \link{diagTab} function to create 2x2 contingency table.}

\item{...}{other arguments to be passed to \link[DescTools:BinomCI]{DescTools::BinomCI}.}

\item{withref}{(\code{logical})\cr weather the comparison test is reference/gold standard or not.}

\item{method}{(\code{string})\cr string specifying the which method to use. Default is \code{wilson}.
Options can be \code{wilson}, \code{wald} and \code{clopper-pearson}, see \link[DescTools:BinomCI]{DescTools::BinomCI}.}

\item{alpha}{(\code{numeric})\cr type-I-risk, \eqn{\alpha}.}

\item{digits}{(\code{integer})\cr the desired number of digits. Default is 4.}
}
\value{
A data frame contains the qualitative diagnostic accuracy criteria with
three columns for estimated value and confidence interval.
\itemize{
\item sens: Sensitivity refers to how often the test is positive when the condition
of interest is present.
\item spec: Specificity refers to how often the test is negative when the condition
of interest is absent.
\item ppv: Positive predictive value refers to the percentage of subjects with
a positive test result who have the target condition.
\item npv: Negative predictive value refers to the percentage of subjects with
a negative test result who do not have the target condition.
\item plr: Positive likelihood ratio refers to the probability of true positive
rate divided by the false negative rate.
\item nlr: Negative likelihood ratio refers to the probability of false positive
rate divided by the true negative rate.
\item ppa: Positive percent agreement, equals to sensitivity when the candidate method
is evaluated by comparison with a comparative method, not reference/gold standard.
\item npa: Negative percent agreement, equals to specificity when the candidate method
is evaluated by comparison with a comparative method, not reference/gold standard.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Provides a concise summary of the content of \code{\link{MCTab}} objects. Computes
sensitivity, specificity, positive and negative predictive values and positive
and negative likelihood ratios for a diagnostic test with reference/gold standard.
Computes positive and negative percent agreement, and overall percent agreement
when the new test is evaluated by comparison to a non-reference standard.
}
\examples{
tab <- qualData \%>\%
  diagTab(
    formula = ~ CandidateN + ComparativeN,
    levels = c(1, 0)
  )
getAccuracy(tab, method = "wilson")
getAccuracy(tab, method = "wilson", withref = FALSE)
}
