% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mcradds-package}
\alias{mcradds}
\alias{mcradds-package}
\title{\code{mcIVD} Package}
\description{
\code{mcIVD} facilitates pre-processing, analyzing, and reporting of IVD data.
}
\author{
\strong{Maintainer}: Kai Gu \email{gukai1212@163.com}

}
