% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mcradds-package}
\alias{mcradds}
\alias{mcradds-package}
\title{\code{mcradds} Package}
\description{
\code{mcradds} Processing and analyzing of In Vitro Diagnostic Data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kaigu1990/mcradds/}
  \item \url{https://kaigu1990.github.io/mcradds/}
  \item Report bugs at \url{https://github.com/kaigu1990/mcradds/issues}
}

}
\author{
\strong{Maintainer}: Kai Gu \email{gukai1212@163.com}

}
