% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_class.R
\docType{class}
\name{MCTab-class}
\alias{MCTab-class}
\alias{MCTab}
\title{MCTab Class}
\usage{
MCTab(data, tab, levels)
}
\arguments{
\item{data}{(\code{data.frame})\cr original data set.}

\item{tab}{(\code{table})\cr \code{table} class converted from \code{\link[=table]{table()}} to display 2x2 contingency table.}

\item{levels}{(\code{character})\cr levels of measurements.}
}
\value{
An object of class \code{MCTab}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{MCTab} class serves as the store for 2x2 contingency table
}
\section{Slots}{

\describe{
\item{\code{data}}{data}

\item{\code{tab}}{candidate}

\item{\code{levels}}{levels}
}}

