% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{platelet}
\alias{platelet}
\title{Example \code{platelet} Data}
\format{
A \link{platelet} contains 120 observations and 3 variables.
\describe{
\item{sample}{Sample id}
\item{Comparative}{Measurement from comparative analyzer}
\item{Candidate}{Measurement from candidate analyzer}
}
}
\source{
CLSI-EP09 A3 Appendix H, Table H2 is cited in this data set.
}
\usage{
platelet
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This example \code{\link{platelet}} can be used to create a data set comparing
Platelet results from two analyzers in cells/μL.
}
\seealso{
\link[mcr:creatinine]{mcr::creatinine} which contains data with with serum and plasma
creatinin measurements in mg/dL for each sample.
}
\keyword{datasets}
